Fix shitty code
Refactor the shitty code that my team-mate wrote using AI (that too fucking ChatGPT). fucking idiots
