+++ MomCare/Controllers/MyPlanControllers/Exercise/BreathingPlayerViewController.swift
@@ -4,7 +4,11 @@ class BreathingPlayerViewController: UIViewController {
 
     // MARK: Internal
 
+    // MARK: - IBOutlets
+
     @IBOutlet var totalBreatingDuration: UILabel!
+
+    // MARK: - Public Properties
 
     var exerciseProgressViewController: ExerciseProgressViewController?
     var remainingMinSec: Double = 0.0
@@ -12,17 +16,11 @@ class BreathingPlayerViewController: UIViewController {
 
     override func viewDidLoad() {
         super.viewDidLoad()
-        setupAnimatedGradientBackground()
-        setupCircleLayers()
-        setupInstructionLabel()
-        setupTimerLabel()
-        setupAssuringMessageLabel()
-        setupControlButtons()
-
-        if let completionDuration: Double? = Utils.get(fromKey: "BreathingCompletionDuration", withDefaultValue: 0.0) {
-            totalBreathingTime -= completionDuration ?? 0
-        }
+        setupUI()
+        setupAnimationManager()
     }
+
+    // MARK: - Public Methods
 
     func updateGradientBackground() {
         let gradientLayer = CAGradientLayer()
@@ -66,98 +64,49 @@ class BreathingPlayerViewController: UIViewController {
         completedPercentage = (completedTime / totalBreathingTime * 100)
     }
 
+    nonisolated func hideTimer() {
+        DispatchQueue.main.async {
+            self.timer?.invalidate()
+            self.timer = nil
+            self.timerLabel.isHidden = true
+            self.timerLabel.text = ""
+        }
+    }
+
     // MARK: Private
 
-    private enum PetalAnimationPhase { case none, expanding, collapsing }
+    // MARK: - Private Properties
 
     private enum PlayerState { case ready, playing, paused, finished }
 
-    private var animatedGradientLayer: CAGradientLayer?
-    private var gradientAnimation: CABasicAnimation?
-
+    private let animationManager: BreathingAnimationManager = .init()
     private let assuringMessageLabel: UILabel = .init()
-    private var circlesContainer: CALayer = .init()
-    private var circleLayers: [CAShapeLayer] = []
-    private var isInhaling = true
     private let instructionLabel: UILabel = .init()
     private let timerLabel: UILabel = .init()
     private var timer: Timer?
     private var currentCount = 0
-    private var breathingCycles = 0
-
-    private var petalAnimationPhase: PetalAnimationPhase = .none
-    private var petalAnimationStartTime: TimeInterval = 0
-    private var petalAnimationDuration: TimeInterval = 0
-    private var petalAnimationRemaining: TimeInterval = 0
-    private var petalTargetPositions: [CGPoint] = []
-    private var petalPausedPositions: [CGPoint] = []
     private var playerState: PlayerState = .ready
     private var startPauseButton: UIButton!
     private var stopButton: UIButton!
     private var exerciseTimer: Timer?
     private var secondsElapsed = 0
 
-    private var nextStateWorkItem: DispatchWorkItem?
-    private var animationPhaseStartTime: TimeInterval = 0
-    private var timeRemainingForPhase: TimeInterval = 0
-
-    private let numberOfPetals = 6
-    private let circleSize: CGFloat = 100
-    private let animationDuration: TimeInterval = 4.0
-    private let spreadDistance: CGFloat = 60
-    private let textAnimationDuration: TimeInterval = 0.5 //
+    private let textAnimationDuration: TimeInterval = 0.5
     private var totalBreathingTime: TimeInterval = 600
-    private let petalColors: [UIColor] = [
-        UIColor(hex: "#bfaee0"),
-        UIColor(hex: "#f7d6e0"),
-        UIColor(hex: "#e6d6f7"),
-        UIColor(hex: "#f7d6ec"),
-        UIColor(hex: "#ffd6d6"),
-        UIColor(hex: "#e3c6f7")
-    ]
-    private let centerColor: UIColor = .init(hex: "#fff6f0")
 
-    private func setupAnimatedGradientBackground() {
-        let gradientLayer = CAGradientLayer()
-        gradientLayer.frame = view.bounds
+    // MARK: - Private Methods
 
-        let color1 = UIColor(hex: "#1e0d31")
-        let color2 = UIColor(hex: "#3a2766")
-        let color3 = UIColor(hex: "#6e4fa3")
-        let color4 = UIColor(hex: "#bfaee0")
-        let color5 = UIColor(hex: "#f7d6e0")
-        gradientLayer.colors = [
-            color1.cgColor,
-            color2.cgColor,
-            color3.cgColor,
-            color4.cgColor,
-            color5.cgColor
-        ]
-        gradientLayer.startPoint = CGPoint(x: 0.5, y: 0.0)
-        gradientLayer.endPoint = CGPoint(x: 0.5, y: 1.0)
-        view.layer.insertSublayer(gradientLayer, at: 0)
-        animatedGradientLayer = gradientLayer
-        animateGradientColors()
+    private func setupUI() {
+        setupAssuringMessageLabel()
+        setupInstructionLabel()
+        setupTimerLabel()
+        setupControlButtons()
     }
 
-    private func animateGradientColors() {
-        guard let gradientLayer = animatedGradientLayer else { return }
-        let fromColors = gradientLayer.colors
-        let toColors: [CGColor] = [
-            UIColor(hex: "#2a1a4d").cgColor,
-            UIColor(hex: "#4b3577").cgColor,
-            UIColor(hex: "#8c6fc9").cgColor,
-            UIColor(hex: "#e3c6f7").cgColor,
-            UIColor(hex: "#f7d6e0").cgColor
-        ]
-        let animation = CABasicAnimation(keyPath: "colors")
-        animation.fromValue = fromColors
-        animation.toValue = toColors
-        animation.duration = 6.0
-        animation.autoreverses = true
-        animation.repeatCount = .infinity
-        animation.isRemovedOnCompletion = false
-        gradientLayer.add(animation, forKey: "colorChange")
+    private func setupAnimationManager() {
+        animationManager.delegate = self
+        animationManager.setupAnimatedGradientBackground(in: view)
+        animationManager.setupCircleLayers(in: view)
     }
 
     private func setupAssuringMessageLabel() {
@@ -209,7 +158,6 @@ class BreathingPlayerViewController: UIViewController {
         }
 
         UIView.animate(withDuration: textAnimationDuration, delay: 0, options: .curveLinear) {} completion: { _ in
-
             self.instructionLabel.text = newText
             self.instructionLabel.transform = .identity
             self.instructionLabel.alpha = 1
@@ -234,7 +182,6 @@ class BreathingPlayerViewController: UIViewController {
     }
 
     private func startTimer(from initialCount: Int = 4) {
-
         timer?.invalidate()
         currentCount = initialCount
 
@@ -254,187 +201,6 @@ class BreathingPlayerViewController: UIViewController {
             timerLabel.text = "\(currentCount)"
         }
         currentCount -= 1
-    }
-
-    private func setupCircleLayers() {
-
-        circlesContainer.removeFromSuperlayer()
-        circleLayers.removeAll()
-
-        let containerSize = circleSize + (spreadDistance * 2)
-        circlesContainer = CALayer()
-        circlesContainer.frame = CGRect(x: 0, y: 0, width: containerSize, height: containerSize)
-        circlesContainer.position = view.center
-        view.layer.addSublayer(circlesContainer)
-
-        for i in 0...numberOfPetals {
-            let circle = createCircleLayer(index: i)
-
-            circle.position = CGPoint(x: circlesContainer.bounds.midX, y: circlesContainer.bounds.midY)
-            circlesContainer.addSublayer(circle)
-            circleLayers.append(circle)
-        }
-    }
-
-    private func hideTimer() {
-        timer?.invalidate()
-        timer = nil
-        timerLabel.isHidden = true
-        timerLabel.text = ""
-    }
-
-    private func createCircleLayer(index: Int) -> CAShapeLayer {
-        let layer = CAShapeLayer()
-        layer.frame = CGRect(x: -circleSize/2, y: -circleSize/2, width: circleSize, height: circleSize)
-        let circlePath = UIBezierPath(
-            arcCenter: CGPoint(x: circleSize/2, y: circleSize/2),
-            radius: circleSize/2,
-            startAngle: 0,
-            endAngle: 2 * .pi,
-            clockwise: true
-        )
-        layer.path = circlePath.cgPath
-        layer.shadowPath = circlePath.cgPath
-
-        if index == 0 {
-            layer.fillColor = centerColor.withAlphaComponent(0.85).cgColor
-        } else {
-            let color = petalColors[(index-1) % petalColors.count]
-            layer.fillColor = color.withAlphaComponent(0.7).cgColor
-            layer.shadowColor = color.withAlphaComponent(0.7).cgColor
-            layer.shadowOpacity = 0.5
-            layer.shadowRadius = 16
-            layer.shadowOffset = CGSize(width: 0, height: 0)
-        }
-        return layer
-    }
-
-    private func startBreathingAnimation() {
-        animateBreathCycle()
-    }
-
-    private func animateBreathCycle() {
-        if playerState != .playing { return }
-
-        if isInhaling {
-            animateInstructionChange(to: "Inhale")
-            hideTimer()
-            animateFlowerFormation {
-                if self.playerState != .playing { return }
-                self.animateInstructionChange(to: "Hold")
-                self.startTimer()
-                self.scheduleNextState(after: self.animationDuration)
-            }
-        } else {
-            currentCount = 4
-            animateInstructionChange(to: "Exhale")
-            hideTimer()
-            animateFlowerCollapse {
-                if self.playerState != .playing { return }
-                self.breathingCycles += 1
-                self.isInhaling = true
-
-                self.animateBreathCycle()
-            }
-        }
-    }
-
-    private func scheduleNextState(after delay: TimeInterval) {
-        animationPhaseStartTime = CACurrentMediaTime()
-        nextStateWorkItem = DispatchWorkItem { [weak self] in
-            guard let self, playerState == .playing else { return }
-
-            isInhaling = false
-            animateBreathCycle()
-        }
-        DispatchQueue.main.asyncAfter(deadline: .now() + delay, execute: nextStateWorkItem!)
-    }
-
-    private func animateFlowerFormation(completion: @escaping () -> Void) {
-        petalAnimationPhase = .expanding
-        petalAnimationStartTime = CACurrentMediaTime()
-        petalAnimationDuration = animationDuration
-        petalAnimationRemaining = animationDuration
-        petalTargetPositions = []
-        petalPausedPositions = []
-        CATransaction.begin()
-        CATransaction.setCompletionBlock {
-            self.pulsePetals()
-            completion()
-        }
-        for (index, circle) in circleLayers.enumerated() {
-            if index == 0 { continue }
-            let angle = (2.0 * .pi * CGFloat(index - 1)) / CGFloat(numberOfPetals)
-            let destinationX = circlesContainer.bounds.midX + cos(angle) * spreadDistance
-            let destinationY = circlesContainer.bounds.midY + sin(angle) * spreadDistance
-            let destination = CGPoint(x: destinationX, y: destinationY)
-            petalTargetPositions.append(destination)
-            let animation = CABasicAnimation(keyPath: "position")
-            animation.fromValue = circle.presentation()?.position ?? circle.position
-            animation.toValue = destination
-            animation.duration = animationDuration
-            animation.timingFunction = CAMediaTimingFunction(name: .easeInEaseOut)
-            animation.fillMode = .forwards
-            animation.isRemovedOnCompletion = false
-            circle.add(animation, forKey: "position")
-        }
-        CATransaction.commit()
-    }
-
-    private func pulsePetals() {
-        for (index, circle) in circleLayers.enumerated() {
-            if index == 0 { continue }
-            let pulse = CABasicAnimation(keyPath: "transform.scale")
-            pulse.fromValue = 1.0
-            pulse.toValue = 1.08
-            pulse.duration = 0.7
-            pulse.autoreverses = true
-            pulse.repeatCount = 2
-            pulse.timingFunction = CAMediaTimingFunction(name: .easeInEaseOut)
-            circle.add(pulse, forKey: "pulse")
-        }
-    }
-
-    private func animateFlowerBloomAndShowMessage() {
-
-        for (index, circle) in circleLayers.enumerated() {
-            if index == 0 { continue }
-            let bloom = CABasicAnimation(keyPath: "transform.scale")
-            bloom.fromValue = 1.0
-            bloom.toValue = 1.25
-            bloom.duration = 1.2
-            bloom.timingFunction = CAMediaTimingFunction(name: .easeInEaseOut)
-            bloom.fillMode = .forwards
-            bloom.isRemovedOnCompletion = false
-            circle.add(bloom, forKey: "bloom")
-        }
-
-        DispatchQueue.main.asyncAfter(deadline: .now() + 1.1) {
-            self.showAssuringMessage()
-        }
-    }
-
-    private func animateFlowerCollapse(completion: @escaping () -> Void) {
-        petalAnimationPhase = .collapsing
-        petalAnimationStartTime = CACurrentMediaTime()
-        petalAnimationDuration = animationDuration
-        petalAnimationRemaining = animationDuration
-        petalTargetPositions = Array(repeating: CGPoint(x: circlesContainer.bounds.midX, y: circlesContainer.bounds.midY), count: numberOfPetals)
-        petalPausedPositions = []
-        CATransaction.begin()
-        CATransaction.setCompletionBlock(completion)
-        for (index, circle) in circleLayers.enumerated() {
-            if index == 0 { continue }
-            let animation = CABasicAnimation(keyPath: "position")
-            animation.fromValue = circle.presentation()?.position ?? circle.position
-            animation.toValue = CGPoint(x: circlesContainer.bounds.midX, y: circlesContainer.bounds.midY)
-            animation.duration = animationDuration
-            animation.timingFunction = CAMediaTimingFunction(name: .easeInEaseOut)
-            animation.fillMode = .forwards
-            animation.isRemovedOnCompletion = false
-            circle.add(animation, forKey: "position")
-        }
-        CATransaction.commit()
     }
 
     private func setupControlButtons() {
@@ -476,7 +242,6 @@ class BreathingPlayerViewController: UIViewController {
     @objc private func startPauseButtonTapped() {
         switch playerState {
         case .ready:
-
             playerState = .playing
             startExercise()
             startPauseButton.setTitle("Pause", for: .normal)
@@ -485,19 +250,16 @@ class BreathingPlayerViewController: UIViewController {
             }
 
         case .playing:
-
             playerState = .paused
             pauseExercise()
             startPauseButton.setTitle("Resume", for: .normal)
 
         case .paused:
-
             playerState = .playing
             resumeExercise()
             startPauseButton.setTitle("Pause", for: .normal)
 
         case .finished:
-
             resetExercise()
         }
     }
@@ -517,8 +279,6 @@ class BreathingPlayerViewController: UIViewController {
                 nav.popViewController(animated: true)
             } else {
                 self.dismiss(animated: true) {
-                    Utils.save(forKey: "BreathingCompletionDuration", withValue: self.totalBreathingTime - self.remainingMinSec)
-
                     self.exerciseProgressViewController?.triggerRefresh()
                 }
             }
@@ -528,96 +288,31 @@ class BreathingPlayerViewController: UIViewController {
     }
 
     private func startExercise() {
-
         secondsElapsed = 0
         updateMainTimer()
         startMainTimer()
-        startBreathingAnimation()
+        animationManager.startBreathingAnimation()
     }
 
     private func pauseExercise() {
         if playerState != .paused { return }
 
         exerciseTimer?.invalidate()
-        timer?.invalidate()
-
-        if petalAnimationPhase != .none {
-            let now = CACurrentMediaTime()
-            let elapsed = now - petalAnimationStartTime
-            petalAnimationRemaining = max(0, petalAnimationDuration - elapsed)
-            petalPausedPositions = []
-            for (index, circle) in circleLayers.enumerated() {
-                if index == 0 { continue }
-                let pos = circle.presentation()?.position ?? circle.position
-                petalPausedPositions.append(pos)
-
-                circle.removeAllAnimations()
-                circle.position = pos
-            }
-        }
-
-        nextStateWorkItem?.cancel()
-        let timeElapsed = CACurrentMediaTime() - animationPhaseStartTime
-        timeRemainingForPhase = animationDuration - timeElapsed
-        if timeRemainingForPhase < 0 { timeRemainingForPhase = 0 }
+        animationManager.pauseAnimation()
     }
 
     private func resumeExercise() {
         if playerState != .playing { return }
 
         startMainTimer()
-
-        if petalAnimationPhase != .none && petalAnimationRemaining > 0 && !petalPausedPositions.isEmpty {
-            for (i, circle) in circleLayers.enumerated() {
-                if i == 0 { continue }
-                let from = petalPausedPositions[i-1]
-                let to: CGPoint
-                if petalAnimationPhase == .expanding {
-                    to = petalTargetPositions[i-1]
-                } else {
-                    to = CGPoint(x: circlesContainer.bounds.midX, y: circlesContainer.bounds.midY)
-                }
-                let animation = CABasicAnimation(keyPath: "position")
-                animation.fromValue = from
-                animation.toValue = to
-                animation.duration = petalAnimationRemaining
-                animation.timingFunction = CAMediaTimingFunction(name: .easeInEaseOut)
-                animation.fillMode = .forwards
-                animation.isRemovedOnCompletion = false
-                circle.position = from
-                circle.add(animation, forKey: "position")
-            }
-
-            petalAnimationStartTime = CACurrentMediaTime()
-            petalAnimationDuration = petalAnimationRemaining
-
-            if isInhaling {
-                scheduleNextState(after: timeRemainingForPhase)
-            }
-            petalAnimationRemaining = 0
-            petalPausedPositions = []
-        }
-
-        if !timerLabel.isHidden {
-            startTimer(from: currentCount)
-        }
-
-        if isInhaling && petalAnimationPhase == .none {
-            scheduleNextState(after: timeRemainingForPhase)
-        }
+        animationManager.resumeAnimation()
     }
 
     private func resetExercise() {
         exerciseTimer?.invalidate()
         secondsElapsed = 0
-        breathingCycles = 0
-        isInhaling = true
 
-        circlesContainer.removeAllAnimations()
-        for circle in circleLayers {
-            circle.removeAllAnimations()
-            circle.position = CGPoint(x: circlesContainer.bounds.midX, y: circlesContainer.bounds.midY)
-        }
+        animationManager.resetAnimation()
 
         startPauseButton.setTitle("Start", for: .normal)
 
@@ -654,7 +349,7 @@ class BreathingPlayerViewController: UIViewController {
             totalBreatingDuration.text = "00:00"
             exerciseTimer?.invalidate()
             playerState = .finished
-            animateFlowerBloomAndShowMessage()
+            animationManager.finishExerciseAnimation()
 
             startPauseButton.setTitle("Restart", for: .normal)
 
@@ -676,12 +371,40 @@ class BreathingPlayerViewController: UIViewController {
     }
 
     @objc private func doneButtonTapped() {
-
         if let nav = navigationController {
             nav.popViewController(animated: true)
         } else {
             dismiss(animated: true)
         }
     }
+}
 
+// MARK: - BreathingAnimationDelegate
+
+extension BreathingPlayerViewController: BreathingAnimationDelegate {
+    nonisolated func animationDidCompleteInhale() {
+        // Handle inhale completion if needed
+    }
+
+    nonisolated func animationDidCompleteExhale() {
+        // Handle exhale completion if needed
+    }
+
+    nonisolated func animationDidFinishExercise() {
+        DispatchQueue.main.async {
+            self.showAssuringMessage()
+        }
+    }
+
+    nonisolated func updateInstructionText(_ text: String) {
+        DispatchQueue.main.async {
+            self.animateInstructionChange(to: text)
+        }
+    }
+
+    nonisolated func showTimer(from count: Int) {
+        DispatchQueue.main.async {
+            self.startTimer(from: count)
+        }
+    }
 }
